+++ cmd/shortener/main.go
@@ -10,7 +10,7 @@ import (
 	"net/http"
 	"strings"
 
-	"github.com/achufistov/shortygopher.git/cmd/shortener/config"
+	"github.com/achufistov/shortygopher.git/internal/app/config"
 
 	"github.com/go-chi/chi/v5"
 )
